Breaking out ServiceWithRegistrationsEditForm into another small component so that the parent ServiceEditForm isn't manipulated as much and the library registration component can have its own state.
